[old/new protocol] Move (old) AwakeableIdentifier in restate_types, Introduce new ExternalSignalIdentifier used to complete awakeables created by journals using journal table V2.
