+++ crates/service-protocol/Cargo.toml
@@ -9,12 +9,10 @@ publish = false
 
 [features]
 default = []
-
-awakeable-id = ["dep:base64", "dep:restate-base64-util", "dep:restate-types"]
 codec = ["dep:restate-types", "dep:paste"]
 discovery = ["dep:serde", "dep:serde_json", "dep:bytestring", "dep:regress", "dep:tracing", "dep:codederror", "dep:restate-errors", "dep:http", "dep:http-body-util", "dep:restate-service-client", "dep:restate-types", "dep:tokio"]
 message = ["dep:restate-types", "dep:bytes-utils", "dep:codederror", "dep:restate-errors", "dep:size", "dep:tracing"]
-test-util = ["awakeable-id"]
+test-util = []
 
 [dependencies]
 workspace-hack = { version = "0.1", path = "../../workspace-hack" }
@@ -24,7 +22,6 @@ restate-errors = { workspace = true, optional = true }
 restate-service-client =  { workspace = true, optional = true }
 restate-types = { workspace = true, optional = true }
 
-base64 = { workspace = true, optional = true }
 bytes = { workspace = true }
 bytestring = { workspace = true, optional = true }
 bytes-utils = { workspace = true, optional = true }
