+++ crates/types/src/identifiers.rs
@@ -10,12 +10,14 @@
 
 //! Restate uses many identifiers to uniquely identify its services and entities.
 
-use bytes::Bytes;
+use base64::Engine;
+use bytes::{BufMut, Bytes, BytesMut};
 use bytestring::ByteString;
 use rand::RngCore;
+use serde::{Deserialize, Serialize};
 use sha2::{Digest, Sha256};
 use std::fmt;
-use std::fmt::Formatter;
+use std::fmt::{Display, Formatter};
 use std::hash::Hash;
 use std::mem::size_of;
 use std::str::FromStr;
@@ -28,6 +30,7 @@ use crate::id_util::IdDecoder;
 use crate::id_util::IdEncoder;
 use crate::id_util::IdResourceType;
 use crate::invocation::{InvocationTarget, InvocationTargetType, WorkflowHandlerType};
+use crate::journal_v2::SignalId;
 use crate::time::MillisSinceEpoch;
 
 /// Identifying the leader epoch of a partition processor
@@ -956,6 +959,170 @@ ulid_backed_id!(Subscription @with_resource_id);
 ulid_backed_id!(PartitionProcessorRpcRequest);
 ulid_backed_id!(Snapshot @with_resource_id);
 
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct AwakeableIdentifier {
+    invocation_id: InvocationId,
+    entry_index: EntryIndex,
+}
+
+impl ResourceId for AwakeableIdentifier {
+    const SIZE_IN_BYTES: usize = InvocationId::SIZE_IN_BYTES + size_of::<EntryIndex>();
+    const RESOURCE_TYPE: IdResourceType = IdResourceType::Awakeable;
+    const STRING_CAPACITY_HINT: usize = 0; /* Not needed since encoding is custom */
+
+    /// We use a custom strategy for awakeable identifiers since they need to be encoded as base64
+    /// for wider language support.
+    fn push_contents_to_encoder(&self, encoder: &mut IdEncoder<Self>) {
+        let mut input_buf =
+            BytesMut::with_capacity(size_of::<EncodedInvocationId>() + size_of::<EntryIndex>());
+        input_buf.put_slice(&self.invocation_id.to_bytes());
+        input_buf.put_u32(self.entry_index);
+        let encoded_base64 = restate_base64_util::URL_SAFE.encode(input_buf.freeze());
+        encoder.push_str(encoded_base64);
+    }
+}
+
+impl AwakeableIdentifier {
+    pub fn new(invocation_id: InvocationId, entry_index: EntryIndex) -> Self {
+        Self {
+            invocation_id,
+            entry_index,
+        }
+    }
+
+    pub fn into_inner(self) -> (InvocationId, EntryIndex) {
+        (self.invocation_id, self.entry_index)
+    }
+}
+
+impl FromStr for AwakeableIdentifier {
+    type Err = IdDecodeError;
+
+    fn from_str(input: &str) -> Result<Self, Self::Err> {
+        let decoder = IdDecoder::new(input)?;
+        // Ensure we are decoding the right type
+        if decoder.resource_type != Self::RESOURCE_TYPE {
+            return Err(IdDecodeError::TypeMismatch);
+        }
+        let remaining = decoder.cursor.take_remaining()?;
+
+        let buffer = restate_base64_util::URL_SAFE
+            .decode(remaining)
+            .map_err(|_| IdDecodeError::Codec)?;
+
+        if buffer.len() != size_of::<EncodedInvocationId>() + size_of::<EntryIndex>() {
+            return Err(IdDecodeError::Length);
+        }
+
+        let invocation_id: InvocationId =
+            InvocationId::from_slice(&buffer[..size_of::<EncodedInvocationId>()])?;
+        let entry_index = EntryIndex::from_be_bytes(
+            buffer[size_of::<EncodedInvocationId>()..]
+                .try_into()
+                // Unwrap is safe because we check the size above.
+                .unwrap(),
+        );
+
+        Ok(Self {
+            invocation_id,
+            entry_index,
+        })
+    }
+}
+
+impl Display for AwakeableIdentifier {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let mut encoder = IdEncoder::<Self>::new();
+        self.push_contents_to_encoder(&mut encoder);
+        std::fmt::Display::fmt(&encoder.finalize(), f)
+    }
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct ExternalSignalIdentifier {
+    invocation_id: InvocationId,
+    signal_index: u32,
+}
+
+impl ResourceId for ExternalSignalIdentifier {
+    const SIZE_IN_BYTES: usize = InvocationId::SIZE_IN_BYTES + size_of::<EntryIndex>();
+    const RESOURCE_TYPE: IdResourceType = IdResourceType::Signal;
+    const STRING_CAPACITY_HINT: usize = 0; /* Not needed since encoding is custom */
+
+    /// We use a custom strategy for awakeable identifiers since they need to be encoded as base64
+    /// for wider language support.
+    fn push_contents_to_encoder(&self, encoder: &mut IdEncoder<Self>) {
+        let mut input_buf =
+            BytesMut::with_capacity(size_of::<EncodedInvocationId>() + size_of::<EntryIndex>());
+        input_buf.put_slice(&self.invocation_id.to_bytes());
+        input_buf.put_u32(self.signal_index);
+        let encoded_base64 = restate_base64_util::URL_SAFE.encode(input_buf.freeze());
+        encoder.push_str(encoded_base64);
+    }
+}
+
+impl ExternalSignalIdentifier {
+    pub fn new(invocation_id: InvocationId, signal_index: u32) -> Self {
+        Self {
+            invocation_id,
+            signal_index,
+        }
+    }
+
+    pub fn into_inner(self) -> (InvocationId, SignalId) {
+        (self.invocation_id, SignalId::for_index(self.signal_index))
+    }
+}
+
+impl FromStr for ExternalSignalIdentifier {
+    type Err = IdDecodeError;
+
+    fn from_str(input: &str) -> Result<Self, Self::Err> {
+        let decoder = IdDecoder::new(input)?;
+        // Ensure we are decoding the right type
+        if decoder.resource_type != Self::RESOURCE_TYPE {
+            return Err(IdDecodeError::TypeMismatch);
+        }
+        let remaining = decoder.cursor.take_remaining()?;
+
+        let buffer = restate_base64_util::URL_SAFE
+            .decode(remaining)
+            .map_err(|_| IdDecodeError::Codec)?;
+
+        if buffer.len() != size_of::<EncodedInvocationId>() + size_of::<EntryIndex>() {
+            return Err(IdDecodeError::Length);
+        }
+
+        let invocation_id: InvocationId =
+            InvocationId::from_slice(&buffer[..size_of::<EncodedInvocationId>()])?;
+        let signal_index = u32::from_be_bytes(
+            buffer[size_of::<EncodedInvocationId>()..]
+                .try_into()
+                // Unwrap is safe because we check the size above.
+                .unwrap(),
+        );
+
+        Ok(Self {
+            invocation_id,
+            signal_index,
+        })
+    }
+}
+
+impl Display for ExternalSignalIdentifier {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let mut encoder = IdEncoder::<Self>::new();
+        self.push_contents_to_encoder(&mut encoder);
+        std::fmt::Display::fmt(&encoder.finalize(), f)
+    }
+}
+
+impl WithInvocationId for ExternalSignalIdentifier {
+    fn invocation_id(&self) -> InvocationId {
+        self.invocation_id
+    }
+}
+
 #[cfg(any(test, feature = "test-util"))]
 mod mocks {
     use super::*;
@@ -1161,5 +1328,43 @@ mod tests {
         let b: SubscriptionId = a.to_string().parse().unwrap();
         assert_eq!(a, b);
         assert_eq!(a.to_string(), b.to_string());
+    }
+
+    #[test]
+    fn roundtrip_awakeable_id() {
+        let expected_invocation_id = InvocationId::mock_random();
+        let expected_entry_index = 2_u32;
+
+        let input_str = AwakeableIdentifier {
+            invocation_id: expected_invocation_id,
+            entry_index: expected_entry_index,
+        }
+        .to_string();
+        dbg!(&input_str);
+
+        let actual = AwakeableIdentifier::from_str(&input_str).unwrap();
+        let (actual_invocation_id, actual_entry_index) = actual.into_inner();
+
+        assert_eq!(expected_invocation_id, actual_invocation_id);
+        assert_eq!(expected_entry_index, actual_entry_index);
+    }
+
+    #[test]
+    fn roundtrip_signal_id() {
+        let expected_invocation_id = InvocationId::mock_random();
+        let expected_signal_index = 2_u32;
+
+        let input_str = ExternalSignalIdentifier {
+            invocation_id: expected_invocation_id,
+            signal_index: expected_signal_index,
+        }
+        .to_string();
+        dbg!(&input_str);
+
+        let actual = ExternalSignalIdentifier::from_str(&input_str).unwrap();
+        let (actual_invocation_id, actual_signal_id) = actual.into_inner();
+
+        assert_eq!(expected_invocation_id, actual_invocation_id);
+        assert_eq!(SignalId::for_index(expected_signal_index), actual_signal_id);
     }
 }
